// src/models/VendorList.ts
export interface VendorList {
  id: string;
  codFornecedor: string;
  email: string;
  nomeFantasia: string;
  categoria: string;
  telefone: string;
  cnpj: string;
  grupo: string;
  status: string;
  rua: string;
  bairro: string;
  cidade: string;
  unidadeFederativa: string;
  pais: string;
}

export interface Adress {}